add support for 32 bit pid using the PIDHIGH field. This allows the test suite to see if it is supported. w2k3 doesn't seem to support it.

+++ source/include/cli_context.h
@@ -180,7 +180,7 @@ struct cli_session {
 	uint16 vuid;
 
 	/* default pid for this session */
-	uint16 pid;
+	uint32 pid;
 };
 
 /*
